Update patch set 1
Patch Set 1: Code-Review+1

I tested this by setting Developer options -> Pointer location
and looking at the dx and dy values when I performed a touch. They were usually <4 in the long direction and <3 in the short. Touches almost always register as touches, not drags.
If I set the fuzz to a higher value all touches registered as touches but the sensitivity to moving the finger was less. This I tested with the Markers app (free). With this setting curves were smooth. With a setting of 10 curves were noticeably squared off. 
It would be better if the fuzz_y value could be different for different x values. This would allow us to change just the right side of the screen (portrait). This doesn't appear possible.

Patch-set: 1
Label: Code-Review=+1