Update patch set 1
Patch Set 1:

Thanks for the idea; haven't tried that yet. Out of the dozen or more different tests, I'm beginning to think that we are indeed hitting hardware limitations on these Nooks, possibly related to the high TPUT OPPs, and their voltages. For example, I had it working well for about an hour, then flickers with five or more HWC overlays, followed by frequent reboots until I reverted to the default kernel.

Patch-set: 1